requirements.txt
anyascii==0.3.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
boto3==1.20.22
botocore==1.23.22
certifi==2021.5.30
charset-normalizer==2.0.12
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-crispy-forms==1.12.0
django-dotenv==1.4.2
django-filter==21.1
django-heroku==0.3.1
django-modelcluster==5.2
django-storages==1.12.3
django-taggit==2.0.0
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
html5lib==1.1
idna==3.3
jmespath==0.10.0
l18n==2021.3
openpyxl==3.0.9
Pillow==8.3.2
platformdirs==2.2.0
psycopg2==2.9.2
psycopg2-binary==2.9.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.27.1
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.1
tablib==3.2.0
telepath==0.2
urllib3==1.26.7
virtualenv==20.7.0
virtualenv-clone==0.5.6
wagtail==2.16.1
webencodings==0.5.1
whitenoise==5.3.0
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
